Fix forget not forgetting it did not forget yet.
to clarify: if a player should have been forgotten, the data of its first opening remained and therefore the treasure acted every time as if it was forgotten - since enough time since the first time opened at all passed. But since it should have forgotten that the player has had opened it, so should have been the time of the first opening.